# Add recursion_mode ('immediate' or 'deferred') to Signal and SignalInstance