fixes #292 

This adds the possibility to control the order callbacks during recursive emission events (when one of the callbacks itself emits the same event):

- `"immediate"`: Nested emission events immediately begin a deeper emission loop:
calling all callbacks with the arguments of the nested emitter before
returning to continue the original emission loop.  This is the default
behavior.
- `"deferred"`: Nested emission events are deferred until the current emission
loop is complete.  This is the default behavior.