@tlambert03 2024-03-10
The failing benchmarks represent a slight increase in the time to create a signal instance:


| Change   | Before [f4ac7c3f] <v0.10.0^0>   | After [b17be85c]    |   Ratio | Benchmark (Parameter)                                      |
|----------|---------------------------------|---------------------|---------|------------------------------------------------------------|
| +        | 477±3ns                         | 833±3ns             |    1.74 | benchmarks.CreateSuite.time_create_signal_instance         |
| +        | 4.80±0.05μs                     | 5.87±0.08μs         |    1.22 | benchmarks.EventedSetSuite.time_create_set(10)             |
| +        | 4.85±0.03μs                     | 5.92±0.02μs         |    1.22 | benchmarks.EventedSetWithCallbackSuite.time_create_set(10) |

which is not something I'm worried about here given that it's on the order of ns.

@Czaki, this is ready for final review.  I'll wait for your review on this one before merging. And I intend to follow up in another PR reviewing the exception handling logic slightly